+++ src/serialization/v1.rs
@@ -60,7 +60,10 @@ pub fn serialize_binary(macaroon: &Macaroon) -> Result<Vec<u8>> {
             }
         }
     }
-    serialized.extend(serialize_as_packet(SIGNATURE, macaroon.signature()));
+    serialized.extend(serialize_as_packet(
+        SIGNATURE,
+        macaroon.signature().as_ref(),
+    ));
     Ok(serialized)
 }
 
@@ -287,19 +290,19 @@ mod tests {
         assert!(super::deserialize(b"NDhJe_A==").is_err());
 
         // these failed fuzz testing for this deserializer (V1)
-        assert!(Macaroon::deserialize(&vec![70, 70, 102, 70]).is_err());
+        assert!(Macaroon::deserialize(vec![70, 70, 102, 70]).is_err());
         let tok = URL_SAFE.encode([97, 97, 97, 97, 97, 97, 97, 97, 97, 97, 10]);
-        assert!(Macaroon::deserialize(&tok.as_bytes()).is_err());
+        assert!(Macaroon::deserialize(tok.as_bytes()).is_err());
         let tok = URL_SAFE.encode([
             48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48,
             48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48,
             48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48, 48,
             48, 48, 48, 48, 48, 48, 48, 44, 125, 59, 64,
         ]);
-        assert!(Macaroon::deserialize(&tok.as_bytes()).is_err());
+        assert!(Macaroon::deserialize(tok.as_bytes()).is_err());
         let tok = URL_SAFE.encode([
             48, 48, 49, 48, 49, 48, 52, 48, 48, 48, 48, 48, 48, 48, 48, 32, 126, 10,
         ]);
-        assert!(Macaroon::deserialize(&tok.as_bytes()).is_err());
+        assert!(Macaroon::deserialize(tok.as_bytes()).is_err());
     }
 }